requirements.txt
fastapi
uvicorn
PyJWT
email-validator
